+++ dacite/src/lib.rs
@@ -42,6 +42,7 @@ pub mod khr_xlib_surface;
 pub mod nv_dedicated_allocation;
 pub mod nv_external_memory;
 pub mod nv_external_memory_capabilities;
+pub mod nv_external_memory_win32;
 
 pub use vks::android_wrapper;
 pub use vks::mir_wrapper;